% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Linarius.R
\name{allele.count}
\alias{allele.count}
\title{Function to count allele presence and absence of dominant markers according to ploidy levels}
\usage{
allele.count(xx, ploidy = 2)
}
\arguments{
\item{xx}{A binary datafram of genotypes, individuals as row and alleles as column}

\item{ploidy}{interger or vector, ploidy level of the population or individuals}
}
\description{
This function aims at counting allele presence and absence in a ballance population, this considering the ploidy level.
}
\note{
There is a little biais on small sized sample.
}
\examples{
head(Betula)
allele.count(Betula,code$Ploidy)
}

