% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Linarius.R
\name{allele.frec}
\alias{allele.frec}
\title{Calculation of Allele hetherozygocy with Dominant markers and mixed ploidy levels}
\usage{
allele.frec(xx, ploidy = 2)
}
\arguments{
\item{xx}{A binary datafram of genotypes, individuals as row and alleles as column}

\item{ploidy}{interger or vector, ploidy level of the population or individuals}
}
\description{
This function aims at elaluating allele hetherozygocy in a ballance population, this considering the ploidy level.  Then It calculated an weighted avarage
}
\examples{
head(Betula)
allele.frec(Betula,code$Ploidy)
}

