% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linarius.R
\name{allele.hetero}
\alias{allele.hetero}
\title{Calculation of Allele frequency with Dominant markers and mixed ploidy levels}
\usage{
allele.hetero(xx, ploidy = 2)
}
\arguments{
\item{xx}{A binary datafram of genotypes, individuals as row and alleles as column}

\item{ploidy}{Interger or vector, ploidy level of the population or individuals}
}
\description{
This function aims at evaluating allele frequency in a ballance population, this considering the ploidy level. Then It calculated an weighted avarage
}
\note{
There is a little biais on small sized sample.
}
\examples{
data(Birch)
allele.hetero(Betula,code$Ploidy)

}

