% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapTools.R
\name{boot.reshape.coord}
\alias{boot.reshape.coord}
\title{Generate non-focal anamorphosis}
\usage{
boot.reshape.coord(xx, dist, nboot = 10)
}
\arguments{
\item{xx}{Coordinates of the points to transform}

\item{dist}{A dist object with the distence to outline by anamorphosis}

\item{nboot}{Times you need for your coffee break in seconds}
}
\description{
This function aims to prepare anamorphosis maps or cartograms.
}

