% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapTools.R
\name{Darcy.export}
\alias{Darcy.export}
\title{Export anamorphosis coordinate to Darcy}
\usage{
Darcy.export(xx, file)
}
\arguments{
\item{xx}{An object generated by one of the function expend.coord() or boot.reshape.coord()}

\item{file}{A text to use in filenames}
}
\description{
Creates files to use in the software Darcy for creating anamorphosis maps. Creates 2 files foo-in.txt to use as source and foo-out.txt to use as image.
}

