% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linarius.R
\name{datagen}
\alias{datagen}
\title{Generate a fake dataset fitting allele frequencies and ploidies.}
\usage{
datagen(frec, ploidy)
}
\arguments{
\item{frec}{Allele frequencies: a vector providing the allele frequency of all alleles, every single frequency have to be between 0 and 1}

\item{ploidy}{Ploidy level of every individuals: a vector providing ploidy level of every single individual. Normally integers.}
}
\description{
This function generates a random binary datafram with in row: simulated individuals, and in colomn: alleles. 
The aim of this function is to test hypothesis, and shall not be used for publishing imaginary results (See licence for more information).
}
\note{
Whenever you make something out of this function you have to precise that there are generated data. 
You shall not use this for getting fake experimental data (it constitues a violation of this package licence ).
Whenever you use that function for bad purposes I will find you, get you and kick your ass!
}

