% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapTools.R
\name{dist.coord}
\alias{dist.coord}
\title{Computes Murphy's constant}
\usage{
dist.coord(xx)
}
\arguments{
\item{xx}{A datafram of coordinates}
}
\description{
Computes a internal sum of distence. Resulting number might be used for getting harmonious cartogram
}

