% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linarius.R
\name{distGPS}
\alias{distGPS}
\title{Compute distance between GPS points}
\usage{
distGPS(lat, lon, names, model = "lambert")
}
\arguments{
\item{lat}{A vector with latitudes}

\item{lon}{A vector with longitudes}

\item{names}{A vector with names of position}

\item{model}{A model for computing distence between places, can be any unambigous abreviation of "lambert", "spherical" or "potato"}
}
\description{
Returns a dist object with  distences between location in km
}
\examples{
data(Birch)
distGPS(code$Lat,code$Long,rownames(code))
}

