% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTools.R
\name{harmonize.coord}
\alias{harmonize.coord}
\title{Harmonize anamorphogram with original map}
\usage{
harmonize.coord(xx, yy)
}
\arguments{
\item{xx}{Anamorphosis points coordinate}

\item{yy}{Real points coordinates}
}
\description{
This function aims to prepare anamorphosis maps or cartograms.
This function centers and homogenized the distances and positions of image points according to real points.
}
\note{
Use this function if the difference of position and/or size between image and source is too big. This function is only a cosmetic feature usefull if your public can misinterprate the cartogram.
}

