% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTools.R
\name{heat.points}
\alias{heat.points}
\title{Makes heat map}
\usage{
heat.points(XX, coord, MAX = max(XX), loc = c(0, 0), length = 1000,
  unit = "", pch = 16, cex = par("cex"), digits = 0, offset = 0)
}
\arguments{
\item{XX}{something}

\item{MAX}{I guess it was maximum}

\item{loc}{Position of scalebare (longitude and latitude)}

\item{length}{Length of scalebare}

\item{unit}{Unit of scalebare}
}
\description{
Plot color points according to an alternate variable.
}
\examples{
require(sp)
data(Kenya) 
plot(Geotermal.Area)
plot(Hells.NP,add=T,col="grey98") 
heat.points(Kenya$Asconc, coord ,loc=c(205000,9896000),offset=0.1,length=3000)

}

