% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapTools.R
\name{heat.values}
\alias{heat.values}
\title{Returns values of color of points for heat-map in RGB system}
\usage{
heat.values(XX, MAX = NULL)
}
\arguments{
\item{XX}{Vector of the variable to color}

\item{MAX}{Value supposed to be the hottest. If Nothing is precised, the maximum value of the dataset is chosen.}
}
\description{
Returns values of color of points for heat-map in RGB system according to the high of the value provided.
}
\examples{
data(Kenya) 
heat.values(Kenya$Asconc)

}

