% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapTools.R
\name{RGB.pop}
\alias{RGB.pop}
\title{Transform genotype in RGB colors}
\usage{
RGB.pop(XX, pop, red, green, blue)
}
\arguments{
\item{XX}{A binary datafram of genotypes, individuals as row and alleles as column}

\item{pop}{A vector informing of population every sample belongs to}

\item{red,green,blue}{A vector of alleles to put in a single color group}
}
\description{
This function aims to create a tri-polar heatmap.
a tripolar heat map aims to indicate relatedness between populations.
}
\examples{
require(sp)
data(Leymus)
RGB.pop(Leymus,pop=Leypop,red=c("GreenExtra.160","GreenExtra.105","GreenExtra.176","Green1.188"),green=c("BlueExtra.127","Blue.add.128","BlueExtra.222","Blue.add.129","BlueExtra.182","BlueExtra.131") ,blue=c("Green1.120","GreenExtra.136","Green1.193","Green1.169","Green1.133"))->col.RGB
plot(west)
points(locations,col=rgb(col.RGB, max = 255) ,pch=16,cex=2)
}

