% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapTools.R
\name{spTransform.coord}
\alias{spTransform.coord}
\title{Prepare projected points for Darcy export}
\usage{
spTransform.coord(xx, crs,
  crs.origin = " +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0 ")
}
\arguments{
\item{xx}{An object generated by one of the function expend.coord() or boot.reshape.coord()}

\item{crs}{Beetween   "" the projected information to feed the rgdal::CRS function.}

\item{crs.origin}{Advence information on original data system. If you use this parametter, you are probably misusing Linarius, infringementing the licence, read the file Licence.l2k for more information about the risk.}
}
\description{
Darcy needs projected data, this function converts a Linarius results to a projected data format.  This function needs rgdal.
}
\seealso{
rgdal
}

