% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Linarius.R
\name{allele.count}
\alias{allele.count}
\title{Function to count allele presence and absence of dominant markers according to ploidy levels}
\usage{
allele.count(xx, ploidy = 2)
}
\arguments{
\item{xx}{A binary datafram of genotypes, individuals as row and alleles as column}

\item{ploidy}{Interger or vector, ploidy level of the population or individuals}
}
\description{
This function aims at counting allele presence and absence in a ballance population, this considering the ploidy level.
}
\examples{
data(Birch)
allele.count(Betula,code$Ploidy)

}

