% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Linarius.R
\name{dist.pop}
\alias{dist.pop}
\title{Compute genetic distances between populations}
\usage{
dist.pop(xx, pop, method = "reynolds", ploidy = 2)
}
\arguments{
\item{xx}{A binary datafram of genotypes, individuals as row and alleles as column}

\item{pop}{A vector informing of population every sample belongs to}

\item{method}{A method for computing genetic distence between populations, can be any unambigous abreviation of "euclidean", "reynolds", "roger" or "nei"}

\item{ploidy}{interger or vector, ploidy level of the population or individuals}
}
\description{
Returns a dist object with genetic distences between populations. 4 indexes are availlable: "Euclidean", "Reynolds", "Roger" and "Nei".
}
\examples{
data(Betula)
dist.pop(Betula, code$Pop,ploidy=code$Ploidy,method="rey")
}

