% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapTools.R
\name{heat.values}
\alias{heat.values}
\title{Makes heat map}
\usage{
heat.values(XX, MAX = NULL)
}
\arguments{
\item{XX}{something}

\item{MAX}{I guess it was maximum}
}
\description{
Makes heat map
}

