\name{Linarius-package}
\alias{Linarius-package}
\alias{Linarius}
\docType{package}
\title{
Analyse of dominant marker data with mixed ploidy levels. 
}
\description{
Expermental package in order to analyse dominant marker data with mixed ploidy levels. 
\details{
\tabular{ll}{
Package: \tab Linarius\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2014-03-31\cr
License: \tab Non-free ; Non-Copyleft ; Open-Source ; Non-modifyable ;  Purchasse for free \cr
}
Use this package for analysis of dominant data (AFLP; ISSR; IRAP; REMAP; iPBS; etc… ) 
}
\author{
Benjamin Leduc <gbl1@hi.is> 
~~ The author and/or maintainer of the package ~~
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
