% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MapTools.R
\name{scalebar}
\alias{scalebar}
\title{Draw a scalebar}
\usage{
scalebar(loc, length, unit = "km", degrees = TRUE, cex = par("cex"),
  scale = 1, ...)
}
\arguments{
\item{loc}{Where to put the bar}

\item{unit}{The unit to use on for the scalebare. For autocomputed size, several units are possible: kilometer: "km", mile "mi", nautical mile "nmi", meter "m" or league "li"}

\item{lenght}{how long should it be}
}
\description{
Draw a scalebar on a map,
}
\examples{
require(sp)
data(Leymus)
plot(west)
scalebar(c(-20.33,63.3),8)
}

